requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-firestore==2.13.1
google-cloud-iot==2.8.4
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.16.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-cov==4.1.0